+++ ovbuilder/__init__.py
@@ -5,4 +5,4 @@ Command-line tool for rapidly provisioning VMware virtual machines
 that are ready to join an OpenVox / Puppet fleet.
 """
 
-__version__ = "0.3.0-dev.1"
+__version__ = "0.3.0-dev.2"